Add CoreFoundation framework
