+++ crates/objc2/Cargo.toml
@@ -126,12 +126,14 @@ iai = { version = "0.1", git = "https://github.com/madsmtm/iai", branch = "callg
 static_assertions = "1.1.0"
 memoffset = "0.9.0"
 block2 = { path = "../block2", default-features = false }
+objc2-core-foundation = { path = "../../framework-crates/objc2-core-foundation", default-features = false, features = [
+    "CFCGTypes",
+] }
 objc2-foundation = { path = "../../framework-crates/objc2-foundation", default-features = false, features = [
     "NSArray",
     "NSDate",
     "NSDictionary",
     "NSEnumerator",
-    "NSGeometry",
     "NSKeyValueObserving",
     "NSNotification",
     "NSObject",
