+++ crates/objc2/src/topics/mod.rs
@@ -6,8 +6,8 @@ pub mod about_generated;
 #[doc = include_str!("alternatives.md")]
 pub mod alternatives {}
 #[cfg(not(feature = "gnustep-1-7"))]
-#[doc = include_str!("core_foundation_interop.md")]
-pub mod core_foundation_interop {}
+#[doc = include_str!("crate_interop.md")]
+pub mod crate_interop {}
 #[doc = include_str!("kvo.md")]
 pub mod kvo {}
 #[doc = include_str!("layered_safety.md")]
