+++ framework-crates/objc2-accessibility/Cargo.toml
@@ -19,6 +19,7 @@ workspace = true
 bitflags = { version = "2.5.0", default-features = false, optional = true }
 block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 
 [package.metadata.docs.rs]
@@ -37,21 +38,22 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-foundation/std"]
-alloc = ["block2?/alloc", "objc2/alloc", "objc2-foundation/alloc"]
-bitflags = ["dep:bitflags", "objc2-foundation/bitflags"]
-block2 = ["dep:block2", "objc2-foundation/block2"]
+std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-core-foundation?/std", "objc2-foundation/std"]
+alloc = ["block2?/alloc", "objc2/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc"]
+bitflags = ["dep:bitflags", "objc2-core-foundation?/bitflags", "objc2-foundation/bitflags"]
+block2 = ["dep:block2", "objc2-core-foundation?/block2", "objc2-foundation/block2"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 
 AXAudiograph = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSAttributedString",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
     "objc2-foundation/NSValue",
 ]
 AXBrailleMap = [
-    "objc2-foundation/NSGeometry",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSObject",
 ]
 AXColorUtilities = []
@@ -91,4 +93,5 @@ all = [
     "AXTechnology",
     "bitflags",
     "block2",
+    "objc2-core-foundation",
 ]
