+++ framework-crates/objc2-automator/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [dependencies]
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
 objc2-app-kit = { path = "../objc2-app-kit", version = "0.2.2", default-features = false, optional = true }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 objc2-osa-kit = { path = "../objc2-osa-kit", version = "0.2.2", default-features = false, optional = true }
 
@@ -34,15 +35,16 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "objc2/std", "objc2-app-kit?/std", "objc2-foundation/std", "objc2-osa-kit?/std"]
-alloc = ["objc2/alloc", "objc2-app-kit?/alloc", "objc2-foundation/alloc", "objc2-osa-kit?/alloc"]
+std = ["alloc", "objc2/std", "objc2-app-kit?/std", "objc2-core-foundation?/std", "objc2-foundation/std", "objc2-osa-kit?/std"]
+alloc = ["objc2/alloc", "objc2-app-kit?/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc", "objc2-osa-kit?/alloc"]
 objc2-app-kit = ["dep:objc2-app-kit"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-app-kit?/objc2-core-foundation", "objc2-foundation/objc2-core-foundation", "objc2-osa-kit?/objc2-core-foundation"]
 objc2-osa-kit = ["dep:objc2-osa-kit"]
 
 AMAction = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSDictionary",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSString",
     "objc2-foundation/NSURL",
 ]
@@ -97,6 +99,7 @@ AMWorkflowView = [
     "objc2-app-kit?/NSResponder",
     "objc2-app-kit?/NSUserInterfaceItemIdentification",
     "objc2-app-kit?/NSView",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSCoder",
     "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
@@ -115,5 +118,6 @@ all = [
     "AMWorkspace",
     "AutomatorErrors",
     "objc2-app-kit",
+    "objc2-core-foundation",
     "objc2-osa-kit",
 ]
