+++ framework-crates/objc2-av-kit/Cargo.toml
@@ -19,6 +19,7 @@ workspace = true
 bitflags = { version = "2.5.0", default-features = false, optional = true }
 block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 
 [target.'cfg(target_os = "macos")'.dependencies]
@@ -40,11 +41,12 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-app-kit?/std", "objc2-foundation/std"]
-alloc = ["block2?/alloc", "objc2/alloc", "objc2-app-kit?/alloc", "objc2-foundation/alloc"]
-bitflags = ["dep:bitflags", "objc2-app-kit?/bitflags", "objc2-foundation/bitflags"]
-block2 = ["dep:block2", "objc2-app-kit?/block2", "objc2-foundation/block2"]
+std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-app-kit?/std", "objc2-core-foundation?/std", "objc2-foundation/std"]
+alloc = ["block2?/alloc", "objc2/alloc", "objc2-app-kit?/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc"]
+bitflags = ["dep:bitflags", "objc2-app-kit?/bitflags", "objc2-core-foundation?/bitflags", "objc2-foundation/bitflags"]
+block2 = ["dep:block2", "objc2-app-kit?/block2", "objc2-core-foundation?/block2", "objc2-foundation/block2"]
 objc2-app-kit = ["dep:objc2-app-kit"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-app-kit?/objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 
 AVCaptureView = [
     "objc2-app-kit?/NSAccessibilityProtocols",
@@ -54,6 +56,7 @@ AVCaptureView = [
     "objc2-app-kit?/NSResponder",
     "objc2-app-kit?/NSUserInterfaceItemIdentification",
     "objc2-app-kit?/NSView",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSCoder",
     "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
@@ -79,6 +82,7 @@ AVPlayerView = [
     "objc2-app-kit?/NSResponder",
     "objc2-app-kit?/NSUserInterfaceItemIdentification",
     "objc2-app-kit?/NSView",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSCoder",
     "objc2-foundation/NSError",
@@ -95,6 +99,7 @@ AVRoutePickerView = [
     "objc2-app-kit?/NSResponder",
     "objc2-app-kit?/NSUserInterfaceItemIdentification",
     "objc2-app-kit?/NSView",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSCoder",
     "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
@@ -112,4 +117,5 @@ all = [
     "bitflags",
     "block2",
     "objc2-app-kit",
+    "objc2-core-foundation",
 ]
