+++ framework-crates/objc2-contacts-ui/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [dependencies]
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
 objc2-contacts = { path = "../objc2-contacts", version = "0.2.2", default-features = false, optional = true }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 
 [target.'cfg(target_os = "macos")'.dependencies]
@@ -37,14 +38,16 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "objc2/std", "objc2-app-kit?/std", "objc2-contacts?/std", "objc2-foundation/std"]
-alloc = ["objc2/alloc", "objc2-app-kit?/alloc", "objc2-contacts?/alloc", "objc2-foundation/alloc"]
+std = ["alloc", "objc2/std", "objc2-app-kit?/std", "objc2-contacts?/std", "objc2-core-foundation?/std", "objc2-foundation/std"]
+alloc = ["objc2/alloc", "objc2-app-kit?/alloc", "objc2-contacts?/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc"]
 objc2-app-kit = ["dep:objc2-app-kit"]
 objc2-contacts = ["dep:objc2-contacts"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-app-kit?/objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 
 CNContactPicker = [
     "objc2-app-kit?/NSResponder",
     "objc2-app-kit?/NSView",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSGeometry",
     "objc2-foundation/NSString",
@@ -72,4 +75,5 @@ all = [
     "CNContactViewController",
     "objc2-app-kit",
     "objc2-contacts",
+    "objc2-core-foundation",
 ]
