+++ framework-crates/objc2-core-foundation/Cargo.toml
@@ -1,0 +1,142 @@
+# This file has been automatically generated by `objc2`'s `header-translator`.
+# DO NOT EDIT
+
+[package]
+name = "objc2-core-foundation"
+version.workspace = true
+description = "Bindings to the CoreFoundation framework"
+edition.workspace = true
+rust-version.workspace = true
+keywords = ["cocoa", "apple", "framework", "macos", "ios"]
+categories.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+bitflags = { version = "2.5.0", default-features = false, optional = true }
+block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
+libc = { version = "0.2.80", default-features = false, optional = true }
+objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+
+[package.metadata.docs.rs]
+default-target = "aarch64-apple-darwin"
+features = ["all"]
+targets = [
+    "aarch64-apple-darwin",
+    "x86_64-apple-darwin",
+    "aarch64-apple-ios",
+    "aarch64-apple-tvos",
+    "aarch64-apple-watchos",
+    "aarch64-apple-ios-macabi",
+    "x86_64-unknown-linux-gnu",
+    "i686-unknown-linux-gnu",
+]
+
+[features]
+default = ["std"]
+
+# Currently not possible to turn off, put here for forwards compatibility.
+std = ["alloc", "bitflags?/std", "block2?/std", "libc?/std", "objc2/std"]
+alloc = ["block2?/alloc", "objc2/alloc"]
+bitflags = ["dep:bitflags"]
+block2 = ["dep:block2"]
+libc = ["dep:libc"]
+
+CFArray = []
+CFAttributedString = []
+CFAvailability = []
+CFBag = []
+CFBase = []
+CFBinaryHeap = []
+CFBitVector = []
+CFBundle = []
+CFByteOrder = []
+CFCGTypes = []
+CFCalendar = ["bitflags"]
+CFCharacterSet = []
+CFData = ["bitflags"]
+CFDate = ["bitflags"]
+CFDateFormatter = ["bitflags"]
+CFDictionary = []
+CFError = []
+CFFileDescriptor = []
+CFFileSecurity = ["bitflags"]
+CFLocale = []
+CFMachPort = []
+CFMessagePort = []
+CFNotificationCenter = []
+CFNumber = []
+CFNumberFormatter = ["bitflags"]
+CFPlugIn = []
+CFPreferences = []
+CFPropertyList = ["bitflags"]
+CFRunLoop = ["bitflags"]
+CFSet = []
+CFSocket = ["bitflags"]
+CFStream = ["bitflags"]
+CFString = ["bitflags"]
+CFStringEncodingExt = []
+CFStringTokenizer = ["bitflags"]
+CFTimeZone = []
+CFTree = []
+CFURL = ["bitflags"]
+CFURLAccess = []
+CFURLEnumerator = ["bitflags"]
+CFUUID = []
+CFUserNotification = []
+CFUtilities = []
+CFXMLNode = []
+CFXMLParser = ["bitflags"]
+all = [
+    "CFArray",
+    "CFAttributedString",
+    "CFAvailability",
+    "CFBag",
+    "CFBase",
+    "CFBinaryHeap",
+    "CFBitVector",
+    "CFBundle",
+    "CFByteOrder",
+    "CFCGTypes",
+    "CFCalendar",
+    "CFCharacterSet",
+    "CFData",
+    "CFDate",
+    "CFDateFormatter",
+    "CFDictionary",
+    "CFError",
+    "CFFileDescriptor",
+    "CFFileSecurity",
+    "CFLocale",
+    "CFMachPort",
+    "CFMessagePort",
+    "CFNotificationCenter",
+    "CFNumber",
+    "CFNumberFormatter",
+    "CFPlugIn",
+    "CFPreferences",
+    "CFPropertyList",
+    "CFRunLoop",
+    "CFSet",
+    "CFSocket",
+    "CFStream",
+    "CFString",
+    "CFStringEncodingExt",
+    "CFStringTokenizer",
+    "CFTimeZone",
+    "CFTree",
+    "CFURL",
+    "CFURLAccess",
+    "CFURLEnumerator",
+    "CFUUID",
+    "CFUserNotification",
+    "CFUtilities",
+    "CFXMLNode",
+    "CFXMLParser",
+    "bitflags",
+    "block2",
+    "libc",
+]
