+++ framework-crates/objc2-core-foundation/README.md
@@ -1,0 +1,14 @@
+# `objc2-core-foundation`
+
+[![Latest version](https://badgen.net/crates/v/objc2-core-foundation)](https://crates.io/crates/objc2-core-foundation)
+[![License](https://badgen.net/badge/license/MIT/blue)](../LICENSE.txt)
+[![Documentation](https://docs.rs/objc2-core-foundation/badge.svg)](https://docs.rs/objc2-core-foundation/)
+[![CI](https://github.com/madsmtm/objc2/actions/workflows/ci.yml/badge.svg)](https://github.com/madsmtm/objc2/actions/workflows/ci.yml)
+
+Rust bindings to Apple's framework CoreFoundation.
+
+This README is kept intentionally small to consolidate the documentation, see
+[the Rust docs](https://docs.rs/objc2-core-foundation/) for more details on this crate.
+
+This crate is part of the [`objc2` project](https://github.com/madsmtm/objc2),
+see that for related crates.
