+++ framework-crates/objc2-core-foundation/src/generated
@@ -1,0 +1,1 @@
+../../../generated/CoreFoundation
