+++ framework-crates/objc2-core-image/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [dependencies]
 block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 objc2-metal = { path = "../objc2-metal", version = "0.2.2", default-features = false, optional = true }
 
@@ -36,15 +37,16 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "block2?/std", "objc2/std", "objc2-foundation/std", "objc2-metal?/std"]
-alloc = ["block2?/alloc", "objc2/alloc", "objc2-foundation/alloc", "objc2-metal?/alloc"]
+std = ["alloc", "block2?/std", "objc2/std", "objc2-core-foundation?/std", "objc2-foundation/std", "objc2-metal?/std"]
+alloc = ["block2?/alloc", "objc2/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc", "objc2-metal?/alloc"]
 apple = []
 gnustep-1-7 = ["objc2/gnustep-1-7", "block2?/gnustep-1-7", "objc2-foundation/gnustep-1-7"]
 gnustep-1-8 = ["gnustep-1-7", "objc2/gnustep-1-8", "block2?/gnustep-1-8", "objc2-foundation/gnustep-1-8"]
 gnustep-1-9 = ["gnustep-1-8", "objc2/gnustep-1-9", "block2?/gnustep-1-9", "objc2-foundation/gnustep-1-9"]
 gnustep-2-0 = ["gnustep-1-9", "objc2/gnustep-2-0", "block2?/gnustep-2-0", "objc2-foundation/gnustep-2-0"]
 gnustep-2-1 = ["gnustep-2-0", "objc2/gnustep-2-1", "block2?/gnustep-2-1", "objc2-foundation/gnustep-2-1"]
-block2 = ["dep:block2", "objc2-foundation/block2", "objc2-metal?/block2"]
+block2 = ["dep:block2", "objc2-core-foundation?/block2", "objc2-foundation/block2", "objc2-metal?/block2"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-foundation/objc2-core-foundation", "objc2-metal?/objc2-core-foundation"]
 objc2-metal = ["dep:objc2-metal"]
 
 CIBarcodeDescriptor = [
@@ -53,15 +55,15 @@ CIBarcodeDescriptor = [
     "objc2-foundation/NSUserActivity",
 ]
 CIColor = [
-    "objc2-foundation/NSGeometry",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
 ]
 CIContext = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSData",
     "objc2-foundation/NSDictionary",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSString",
     "objc2-foundation/NSURL",
     "objc2-metal?/MTLCommandQueue",
@@ -73,16 +75,16 @@ CIDetector = [
     "objc2-foundation/NSString",
 ]
 CIFeature = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
 ]
 CIFilter = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSData",
     "objc2-foundation/NSDictionary",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
     "objc2-foundation/NSURL",
@@ -95,14 +97,14 @@ CIFilterGenerator = [
     "objc2-foundation/NSURL",
 ]
 CIFilterShape = [
-    "objc2-foundation/NSGeometry",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSObject",
 ]
 CIImage = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSData",
     "objc2-foundation/NSDictionary",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
     "objc2-foundation/NSURL",
@@ -110,12 +112,12 @@ CIImage = [
     "objc2-metal?/MTLResource",
     "objc2-metal?/MTLTexture",
 ]
-CIImageAccumulator = ["objc2-foundation/NSGeometry"]
+CIImageAccumulator = ["objc2-core-foundation?/CFCGTypes"]
 CIImageProcessor = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSDictionary",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSString",
     "objc2-metal?/MTLAllocation",
     "objc2-metal?/MTLCommandBuffer",
@@ -124,20 +126,20 @@ CIImageProcessor = [
 ]
 CIImageProvider = ["objc2-foundation/NSString"]
 CIKernel = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSData",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSString",
 ]
 CIKernelMetalLib = []
 CIPlugIn = ["objc2-foundation/NSURL"]
 CIPlugInInterface = []
 CIRAWFilter = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSData",
     "objc2-foundation/NSDictionary",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
     "objc2-foundation/NSURL",
@@ -150,9 +152,9 @@ CIRAWFilter_Deprecated = [
     "objc2-foundation/NSURL",
 ]
 CIRenderDestination = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSDate",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-metal?/MTLAllocation",
     "objc2-metal?/MTLCommandBuffer",
     "objc2-metal?/MTLPixelFormat",
@@ -160,13 +162,13 @@ CIRenderDestination = [
     "objc2-metal?/MTLTexture",
 ]
 CISampler = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSDictionary",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
 ]
 CIVector = [
-    "objc2-foundation/NSGeometry",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
 ]
@@ -196,5 +198,6 @@ all = [
     "CIVector",
     "CoreImageDefines",
     "block2",
+    "objc2-core-foundation",
     "objc2-metal",
 ]
