+++ framework-crates/objc2-file-provider/Cargo.toml
@@ -19,6 +19,7 @@ workspace = true
 bitflags = { version = "2.5.0", default-features = false, optional = true }
 block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 objc2-uniform-type-identifiers = { path = "../objc2-uniform-type-identifiers", version = "0.2.2", default-features = false, optional = true }
 
@@ -35,10 +36,11 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-foundation/std", "objc2-uniform-type-identifiers?/std"]
-alloc = ["block2?/alloc", "objc2/alloc", "objc2-foundation/alloc", "objc2-uniform-type-identifiers?/alloc"]
-bitflags = ["dep:bitflags", "objc2-foundation/bitflags"]
-block2 = ["dep:block2", "objc2-foundation/block2", "objc2-uniform-type-identifiers?/block2"]
+std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-core-foundation?/std", "objc2-foundation/std", "objc2-uniform-type-identifiers?/std"]
+alloc = ["block2?/alloc", "objc2/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc", "objc2-uniform-type-identifiers?/alloc"]
+bitflags = ["dep:bitflags", "objc2-core-foundation?/bitflags", "objc2-foundation/bitflags"]
+block2 = ["dep:block2", "objc2-core-foundation?/block2", "objc2-foundation/block2", "objc2-uniform-type-identifiers?/block2"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 objc2-uniform-type-identifiers = ["dep:objc2-uniform-type-identifiers"]
 
 NSFileProviderActions = [
@@ -111,11 +113,11 @@ NSFileProviderManager = [
 NSFileProviderModifyItemOptions = ["bitflags"]
 NSFileProviderReplicatedExtension = [
     "bitflags",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSData",
     "objc2-foundation/NSDictionary",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSProgress",
     "objc2-foundation/NSRange",
     "objc2-foundation/NSString",
@@ -136,10 +138,10 @@ NSFileProviderTesting = [
     "objc2-foundation/NSString",
 ]
 NSFileProviderThumbnailing = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSData",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSProgress",
     "objc2-foundation/NSString",
 ]
@@ -162,5 +164,6 @@ all = [
     "NSFileProviderThumbnailing",
     "bitflags",
     "block2",
+    "objc2-core-foundation",
     "objc2-uniform-type-identifiers",
 ]
