+++ framework-crates/objc2-foundation/Cargo.toml
@@ -20,6 +20,7 @@ bitflags = { version = "2.5.0", default-features = false, optional = true }
 block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
 libc = { version = "0.2.80", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 dispatch = { version = "0.2.0", optional = true }
 
 [package.metadata.docs.rs]
@@ -40,8 +41,8 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "bitflags?/std", "block2?/std", "libc?/std", "objc2/std"]
-alloc = ["block2?/alloc", "objc2/alloc"]
+std = ["alloc", "bitflags?/std", "block2?/std", "libc?/std", "objc2/std", "objc2-core-foundation?/std"]
+alloc = ["block2?/alloc", "objc2/alloc", "objc2-core-foundation?/alloc"]
 apple = []
 gnustep-1-7 = ["objc2/gnustep-1-7", "block2?/gnustep-1-7"]
 gnustep-1-8 = ["gnustep-1-7", "objc2/gnustep-1-8", "block2?/gnustep-1-8"]
@@ -51,13 +52,14 @@ gnustep-2-1 = ["gnustep-2-0", "objc2/gnustep-2-1", "block2?/gnustep-2-1"]
 unstable-static-nsstring = []
 dispatch = ["dep:dispatch"]
 unstable-mutation-return-null = ["NSNull"]
-bitflags = ["dep:bitflags"]
-block2 = ["dep:block2"]
-libc = ["dep:libc"]
+bitflags = ["dep:bitflags", "objc2-core-foundation?/bitflags"]
+block2 = ["dep:block2", "objc2-core-foundation?/block2"]
+libc = ["dep:libc", "objc2-core-foundation?/libc"]
+objc2-core-foundation = ["dep:objc2-core-foundation"]
 
 FoundationErrors = []
 FoundationLegacySwiftCompatibility = []
-NSAffineTransform = []
+NSAffineTransform = ["objc2-core-foundation?/CFCGTypes"]
 NSAppleEventDescriptor = ["bitflags"]
 NSAppleEventManager = []
 NSAppleScript = []
@@ -106,7 +108,10 @@ NSFileVersion = ["bitflags"]
 NSFileWrapper = ["bitflags"]
 NSFormatter = []
 NSGarbageCollector = []
-NSGeometry = ["bitflags"]
+NSGeometry = [
+    "bitflags",
+    "objc2-core-foundation?/CFCGTypes",
+]
 NSHFSFileTypes = []
 NSHTTPCookie = []
 NSHTTPCookieStorage = []
@@ -397,6 +402,7 @@ all = [
     "bitflags",
     "block2",
     "libc",
+    "objc2-core-foundation",
 ]
 
 [dev-dependencies]
