+++ framework-crates/objc2-input-method-kit/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [dependencies]
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
 objc2-app-kit = { path = "../objc2-app-kit", version = "0.2.2", default-features = false, optional = true }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 
 [package.metadata.docs.rs]
@@ -32,12 +33,14 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "objc2/std", "objc2-app-kit?/std", "objc2-foundation/std"]
-alloc = ["objc2/alloc", "objc2-app-kit?/alloc", "objc2-foundation/alloc"]
+std = ["alloc", "objc2/std", "objc2-app-kit?/std", "objc2-core-foundation?/std", "objc2-foundation/std"]
+alloc = ["objc2/alloc", "objc2-app-kit?/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc"]
 objc2-app-kit = ["dep:objc2-app-kit"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-app-kit?/objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 
 IMKCandidates = [
     "objc2-app-kit?/NSResponder",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSAttributedString",
     "objc2-foundation/NSCoder",
@@ -48,6 +51,7 @@ IMKCandidates = [
 IMKInputController = [
     "objc2-app-kit?/NSEvent",
     "objc2-app-kit?/NSMenu",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSAttributedString",
     "objc2-foundation/NSDictionary",
@@ -64,4 +68,5 @@ all = [
     "IMKInputController",
     "IMKServer",
     "objc2-app-kit",
+    "objc2-core-foundation",
 ]
