+++ framework-crates/objc2-metal/Cargo.toml
@@ -19,6 +19,7 @@ workspace = true
 bitflags = { version = "2.5.0", default-features = false, optional = true }
 block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 
 [package.metadata.docs.rs]
@@ -36,14 +37,15 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-foundation/std"]
-alloc = ["block2?/alloc", "objc2/alloc", "objc2-foundation/alloc"]
+std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-core-foundation?/std", "objc2-foundation/std"]
+alloc = ["block2?/alloc", "objc2/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc"]
 unstable-private = [
     "objc2-foundation/NSString",
     "objc2-foundation/NSError",
 ]
-bitflags = ["dep:bitflags", "objc2-foundation/bitflags"]
-block2 = ["dep:block2", "objc2-foundation/block2"]
+bitflags = ["dep:bitflags", "objc2-core-foundation?/bitflags", "objc2-foundation/bitflags"]
+block2 = ["dep:block2", "objc2-core-foundation?/block2", "objc2-foundation/block2"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 
 MTLAccelerationStructure = [
     "bitflags",
@@ -89,6 +91,7 @@ MTLCaptureManager = [
 MTLCaptureScope = ["objc2-foundation/NSString"]
 MTLCommandBuffer = [
     "bitflags",
+    "objc2-core-foundation?/CFDate",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSEnumerator",
     "objc2-foundation/NSError",
@@ -138,7 +141,7 @@ MTLDeviceCertification = [
     "objc2-foundation/NSProcessInfo",
     "objc2-foundation/NSString",
 ]
-MTLDrawable = []
+MTLDrawable = ["objc2-core-foundation?/CFDate"]
 MTLDynamicLibrary = [
     "objc2-foundation/NSError",
     "objc2-foundation/NSString",
@@ -328,6 +331,7 @@ all = [
     "MTLVisibleFunctionTable",
     "bitflags",
     "block2",
+    "objc2-core-foundation",
 ]
 
 [[example]]
@@ -339,6 +343,7 @@ required-features = [
     "objc2-app-kit/NSApplication",
     "objc2-app-kit/NSRunningApplication",
     "objc2-app-kit/NSView",
+    "objc2-core-foundation",
     "objc2-foundation/NSError",
     "objc2-foundation/NSString",
     "objc2-foundation/NSDate",
