+++ framework-crates/objc2-osa-kit/Cargo.toml
@@ -19,6 +19,7 @@ workspace = true
 bitflags = { version = "2.5.0", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
 objc2-app-kit = { path = "../objc2-app-kit", version = "0.2.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 
 [package.metadata.docs.rs]
@@ -33,9 +34,10 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "bitflags?/std", "objc2/std", "objc2-app-kit/std", "objc2-foundation/std"]
-alloc = ["objc2/alloc", "objc2-app-kit/alloc", "objc2-foundation/alloc"]
-bitflags = ["dep:bitflags", "objc2-app-kit/bitflags", "objc2-foundation/bitflags"]
+std = ["alloc", "bitflags?/std", "objc2/std", "objc2-app-kit/std", "objc2-core-foundation?/std", "objc2-foundation/std"]
+alloc = ["objc2/alloc", "objc2-app-kit/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc"]
+bitflags = ["dep:bitflags", "objc2-app-kit/bitflags", "objc2-core-foundation?/bitflags", "objc2-foundation/bitflags"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-app-kit/objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 
 OSALanguage = [
     "bitflags",
@@ -88,6 +90,7 @@ OSAScriptView = [
     "objc2-app-kit/NSUserInterfaceItemIdentification",
     "objc2-app-kit/NSUserInterfaceValidation",
     "objc2-app-kit/NSView",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSCoder",
     "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
@@ -100,4 +103,5 @@ all = [
     "OSAScriptController",
     "OSAScriptView",
     "bitflags",
+    "objc2-core-foundation",
 ]
