+++ framework-crates/objc2-photos-ui/Cargo.toml
@@ -19,6 +19,7 @@ workspace = true
 bitflags = { version = "2.5.0", default-features = false, optional = true }
 block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-core-location = { path = "../objc2-core-location", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 objc2-map-kit = { path = "../objc2-map-kit", version = "0.2.2", default-features = false, optional = true }
@@ -45,11 +46,12 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-app-kit?/std", "objc2-core-location?/std", "objc2-foundation/std", "objc2-map-kit?/std", "objc2-photos?/std"]
-alloc = ["block2?/alloc", "objc2/alloc", "objc2-app-kit?/alloc", "objc2-core-location?/alloc", "objc2-foundation/alloc", "objc2-map-kit?/alloc", "objc2-photos?/alloc"]
-bitflags = ["dep:bitflags", "objc2-app-kit?/bitflags", "objc2-foundation/bitflags", "objc2-map-kit?/bitflags", "objc2-photos?/bitflags"]
-block2 = ["dep:block2", "objc2-app-kit?/block2", "objc2-core-location?/block2", "objc2-foundation/block2", "objc2-map-kit?/block2", "objc2-photos?/block2"]
+std = ["alloc", "bitflags?/std", "block2?/std", "objc2/std", "objc2-app-kit?/std", "objc2-core-foundation?/std", "objc2-core-location?/std", "objc2-foundation/std", "objc2-map-kit?/std", "objc2-photos?/std"]
+alloc = ["block2?/alloc", "objc2/alloc", "objc2-app-kit?/alloc", "objc2-core-foundation?/alloc", "objc2-core-location?/alloc", "objc2-foundation/alloc", "objc2-map-kit?/alloc", "objc2-photos?/alloc"]
+bitflags = ["dep:bitflags", "objc2-app-kit?/bitflags", "objc2-core-foundation?/bitflags", "objc2-foundation/bitflags", "objc2-map-kit?/bitflags", "objc2-photos?/bitflags"]
+block2 = ["dep:block2", "objc2-app-kit?/block2", "objc2-core-foundation?/block2", "objc2-core-location?/block2", "objc2-foundation/block2", "objc2-map-kit?/block2", "objc2-photos?/block2"]
 objc2-app-kit = ["dep:objc2-app-kit", "objc2-map-kit?/objc2-app-kit", "objc2-photos?/objc2-app-kit"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-app-kit?/objc2-core-foundation", "objc2-foundation/objc2-core-foundation", "objc2-map-kit?/objc2-core-foundation", "objc2-photos?/objc2-core-foundation"]
 objc2-core-location = ["dep:objc2-core-location", "objc2-map-kit?/objc2-core-location", "objc2-photos?/objc2-core-location"]
 objc2-map-kit = ["dep:objc2-map-kit"]
 objc2-photos = ["dep:objc2-photos"]
@@ -68,6 +70,7 @@ PHLivePhotoView = [
     "objc2-app-kit?/NSResponder",
     "objc2-app-kit?/NSUserInterfaceItemIdentification",
     "objc2-app-kit?/NSView",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSCoder",
     "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
@@ -108,11 +111,11 @@ PHProjectExtensionController = [
 ]
 PHProjectInfo = [
     "objc2-app-kit?/NSColor",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-core-location?/CLLocation",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSAttributedString",
     "objc2-foundation/NSDate",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
     "objc2-map-kit?/MKAnnotation",
@@ -145,6 +148,7 @@ all = [
     "bitflags",
     "block2",
     "objc2-app-kit",
+    "objc2-core-foundation",
     "objc2-core-location",
     "objc2-map-kit",
     "objc2-photos",
