+++ framework-crates/objc2-screen-capture-kit/Cargo.toml
@@ -20,6 +20,7 @@ bitflags = { version = "2.5.0", default-features = false, optional = true }
 block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
 libc = { version = "0.2.80", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 
 [package.metadata.docs.rs]
@@ -34,11 +35,12 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "bitflags?/std", "block2?/std", "libc?/std", "objc2/std", "objc2-foundation/std"]
-alloc = ["block2?/alloc", "objc2/alloc", "objc2-foundation/alloc"]
-bitflags = ["dep:bitflags", "objc2-foundation/bitflags"]
-block2 = ["dep:block2", "objc2-foundation/block2"]
-libc = ["dep:libc", "objc2-foundation/libc"]
+std = ["alloc", "bitflags?/std", "block2?/std", "libc?/std", "objc2/std", "objc2-core-foundation?/std", "objc2-foundation/std"]
+alloc = ["block2?/alloc", "objc2/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc"]
+bitflags = ["dep:bitflags", "objc2-core-foundation?/bitflags", "objc2-foundation/bitflags"]
+block2 = ["dep:block2", "objc2-core-foundation?/block2", "objc2-foundation/block2"]
+libc = ["dep:libc", "objc2-core-foundation?/libc", "objc2-foundation/libc"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 
 SCContentSharingPicker = [
     "bitflags",
@@ -54,15 +56,15 @@ SCRecordingOutput = [
 ]
 SCScreenshotManager = []
 SCShareableContent = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSString",
 ]
 SCStream = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSString",
 ]
 all = [
@@ -75,4 +77,5 @@ all = [
     "bitflags",
     "block2",
     "libc",
+    "objc2-core-foundation",
 ]
