+++ framework-crates/objc2-screen-saver/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [dependencies]
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
 objc2-app-kit = { path = "../objc2-app-kit", version = "0.2.2", default-features = false, optional = true }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
 
 [package.metadata.docs.rs]
@@ -32,9 +33,10 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "objc2/std", "objc2-app-kit?/std", "objc2-foundation/std"]
-alloc = ["objc2/alloc", "objc2-app-kit?/alloc", "objc2-foundation/alloc"]
+std = ["alloc", "objc2/std", "objc2-app-kit?/std", "objc2-core-foundation?/std", "objc2-foundation/std"]
+alloc = ["objc2/alloc", "objc2-app-kit?/alloc", "objc2-core-foundation?/alloc", "objc2-foundation/alloc"]
 objc2-app-kit = ["dep:objc2-app-kit"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-app-kit?/objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 
 ScreenSaverDefaults = [
     "objc2-foundation/NSString",
@@ -50,6 +52,7 @@ ScreenSaverView = [
     "objc2-app-kit?/NSUserInterfaceItemIdentification",
     "objc2-app-kit?/NSView",
     "objc2-app-kit?/NSWindow",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSCoder",
     "objc2-foundation/NSDate",
     "objc2-foundation/NSGeometry",
@@ -59,4 +62,5 @@ all = [
     "ScreenSaverDefaults",
     "ScreenSaverView",
     "objc2-app-kit",
+    "objc2-core-foundation",
 ]
