+++ framework-crates/objc2-vision/Cargo.toml
@@ -18,6 +18,7 @@ workspace = true
 [dependencies]
 block2 = { path = "../../crates/block2", version = "0.5.1", default-features = false, optional = true }
 objc2 = { path = "../../crates/objc2", version = "0.5.2", default-features = false }
+objc2-core-foundation = { path = "../objc2-core-foundation", version = "0.2.2", default-features = false, optional = true }
 objc2-core-image = { path = "../objc2-core-image", version = "0.2.2", default-features = false, optional = true }
 objc2-core-ml = { path = "../objc2-core-ml", version = "0.2.2", default-features = false, optional = true }
 objc2-foundation = { path = "../objc2-foundation", version = "0.2.2", default-features = false }
@@ -37,9 +38,10 @@ targets = [
 default = ["std"]
 
 # Currently not possible to turn off, put here for forwards compatibility.
-std = ["alloc", "block2?/std", "objc2/std", "objc2-core-image?/std", "objc2-core-ml?/std", "objc2-foundation/std"]
-alloc = ["block2?/alloc", "objc2/alloc", "objc2-core-image?/alloc", "objc2-core-ml?/alloc", "objc2-foundation/alloc"]
-block2 = ["dep:block2", "objc2-core-image?/block2", "objc2-core-ml?/block2", "objc2-foundation/block2"]
+std = ["alloc", "block2?/std", "objc2/std", "objc2-core-foundation?/std", "objc2-core-image?/std", "objc2-core-ml?/std", "objc2-foundation/std"]
+alloc = ["block2?/alloc", "objc2/alloc", "objc2-core-foundation?/alloc", "objc2-core-image?/alloc", "objc2-core-ml?/alloc", "objc2-foundation/alloc"]
+block2 = ["dep:block2", "objc2-core-foundation?/block2", "objc2-core-image?/block2", "objc2-core-ml?/block2", "objc2-foundation/block2"]
+objc2-core-foundation = ["dep:objc2-core-foundation", "objc2-core-image?/objc2-core-foundation", "objc2-foundation/objc2-core-foundation"]
 objc2-core-image = ["dep:objc2-core-image"]
 objc2-core-ml = ["dep:objc2-core-ml"]
 
@@ -146,14 +148,14 @@ VNDetectTrajectoriesRequest = [
     "objc2-foundation/NSObject",
 ]
 VNDetectedPoint = [
-    "objc2-foundation/NSGeometry",
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
 ]
 VNError = ["objc2-foundation/NSString"]
 VNFaceLandmarks = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSValue",
 ]
@@ -200,9 +202,9 @@ VNGeneratePersonSegmentationRequest = [
     "objc2-foundation/NSValue",
 ]
 VNGeometry = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSIndexPath",
     "objc2-foundation/NSObject",
 ]
@@ -225,6 +227,7 @@ VNImageRegistrationRequest = [
     "objc2-foundation/NSURL",
 ]
 VNObservation = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-core-image?/CIBarcodeDescriptor",
     "objc2-core-ml?/MLFeatureValue",
     "objc2-core-ml?/MLMultiArray",
@@ -232,7 +235,6 @@ VNObservation = [
     "objc2-foundation/NSData",
     "objc2-foundation/NSDictionary",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSIndexPath",
     "objc2-foundation/NSIndexSet",
     "objc2-foundation/NSObject",
@@ -258,11 +260,11 @@ VNRecognizedPoint3D = [
     "objc2-foundation/NSString",
 ]
 VNRequest = [
+    "objc2-core-foundation?/CFCGTypes",
     "objc2-core-ml?/MLComputeDeviceProtocol",
     "objc2-foundation/NSArray",
     "objc2-foundation/NSDictionary",
     "objc2-foundation/NSError",
-    "objc2-foundation/NSGeometry",
     "objc2-foundation/NSIndexSet",
     "objc2-foundation/NSObject",
     "objc2-foundation/NSString",
@@ -318,7 +320,7 @@ VNTrackingRequest = [
     "objc2-foundation/NSObject",
 ]
 VNTypes = ["objc2-foundation/NSString"]
-VNUtils = ["objc2-foundation/NSGeometry"]
+VNUtils = ["objc2-core-foundation?/CFCGTypes"]
 VNVideoProcessor = [
     "objc2-foundation/NSDictionary",
     "objc2-foundation/NSError",
@@ -380,6 +382,7 @@ all = [
     "VNUtils",
     "VNVideoProcessor",
     "block2",
+    "objc2-core-foundation",
     "objc2-core-image",
     "objc2-core-ml",
 ]
